# Yet another jpeg encoder optimization