… src/ImageSharp/Formats/Jpeg/Components/Encoder/HuffmanSpec.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/EncodeJpeg.cs
… tests/ImageSharp.Tests/Formats/Jpg/HuffmanScanEncoderTests.cs