### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)


1. Jpeg encoder is now faster across all quality levels!
2. Added some tests for internal `HuffmanScanEncoder` API

| Method | Quality | Mean | Error | StdDev | Ratio |
| ------------------- | -------- | ---------:| ---------:| ---------:| ------:|
| 4:2:0 master | 75 | 29.09 ms | 0.264 ms | 0.234 ms | 1.00 |
| **4:2:0 pr** | 75 | 26.30 ms | 0.131 ms | 0.109 ms | **0.90** |
| 4:2:0 master | 90 | 35.84 ms | 0.163 ms | 0.144 ms | 1.00 |
| **4:2:0 pr**     |      90 | 33.56 ms | 0.237 ms | 0.222 ms |  **0.93** |
| 4:2:0 master | 100 | 42.11 ms | 0.255 ms | 0.226 ms | 1.00 |
| **4:2:0 pr** | 100 | 40.23 ms | 0.225 ms | 0.277 ms | **0.95** |
|                    |         |          |          |          |       |
| 4:4:4 master | 75 | 42.11 ms | 0.206 ms | 0.172 ms | 1.00 |
| **4:4:4 pr** | 75 | 36.70 ms | 0.303 ms | 0.269 ms | **0.87** |
| 4:4:4 master | 90 | 50.12 ms | 0.278 ms | 0.260 ms | 1.00 |
| **4:4:4 pr** | 90 | 44.82 ms | 0.250 ms | 0.234 ms | **0.89** |
| 4:4:4 master | 100 | 69.88 ms | 0.594 ms | 0.496 ms | 1.00 |
| **4:4:4 pr** | 100 | 63.35 ms | 0.486 ms | 0.431 ms | **0.91** |

Almost beating `System.Drawing`
|                      Method | Quality |     Mean |    Error |   StdDev | Ratio |
|---------------------------- |-------- |---------:|---------:|---------:|------:|
| 'System.Drawing Jpeg 4:2:0' |      75 | 29.41 ms | 0.108 ms | 0.096 ms |  1.00 |
|     'ImageSharp Jpeg 4:2:0' |      75 | 26.30 ms | 0.131 ms | 0.109 ms |  0.89 |
|     'ImageSharp Jpeg 4:4:4' |      75 | 36.70 ms | 0.303 ms | 0.269 ms |  1.25 |
|                             |         |          |          |          |       |
| 'System.Drawing Jpeg 4:2:0' |      90 | 32.67 ms | 0.226 ms | 0.211 ms |  1.00 |
|     'ImageSharp Jpeg 4:2:0' |      90 | 33.56 ms | 0.237 ms | 0.222 ms |  1.03 |
|     'ImageSharp Jpeg 4:4:4' |      90 | 44.82 ms | 0.250 ms | 0.234 ms |  1.37 |
|                             |         |          |          |          |       |
| 'System.Drawing Jpeg 4:2:0' |     100 | 39.06 ms | 0.233 ms | 0.218 ms |  1.00 |
|     'ImageSharp Jpeg 4:2:0' |     100 | 40.23 ms | 0.225 ms | 0.277 ms |  1.03 |
|     'ImageSharp Jpeg 4:4:4' |     100 | 63.35 ms | 0.486 ms | 0.431 ms |  1.62 |